+++ app/Providers/AuthServiceProvider.php
@@ -2,13 +2,15 @@
 
 namespace App\Providers;
 
+use App\User;
 use Laravel\Horizon\Horizon;
+use App\Policies\UserPolicy;
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
 
 class AuthServiceProvider extends ServiceProvider
 {
     protected $policies = [
-        'App\Model' => 'App\Policies\ModelPolicy',
+        User::class => UserPolicy::class,
     ];
 
     public function boot()